% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{auc_calculate_withstage}
\alias{auc_calculate_withstage}
\title{Calculate auroc with stage info}
\usage{
auc_calculate_withstage(sce, pretrained_model, query_species, highlight)
}
\arguments{
\item{sce}{a single cell experiment object}

\item{pretrained_model}{a pretrained_model list}

\item{query_species}{options: c('Hs','Mm')}
}
\value{
correlation bewtween query and ref celltypes
}
\description{
Calculate auroc with stage info
}
