% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Cell_qc}
\alias{Cell_qc}
\title{Cell qc}
\usage{
Cell_qc(expression_profile, query_species, metadata, downsample = 20000)
}
\arguments{
\item{expression_profile}{a gene expression matrix}

\item{query_species}{options: c('Hs','Mm')}

\item{metadata}{meta data containing a \code{cell-type} column}

\item{downsample}{set number of cells
default: 20000}
}
\value{
a seurat object of cells passing qulity control
}
\description{
Cell qc
}
