% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{design_matrix}
\alias{design_matrix}
\title{Transform a vector with cell_type labels into a binary matrix}
\usage{
design_matrix(cell_type)
}
\arguments{
\item{cell_type}{ref celltype info}
}
\value{
a binary matix added celltype info
}
\description{
Transform a vector with cell_type labels into a binary matrix
}
