% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_transfer}
\alias{gene_transfer}
\title{gene transfer data}
\format{
A data frame with 2 variables
}
\usage{
gene_transfer
}
\description{
The corresponding of mouse gene and human gene
}
\concept{data}
\keyword{datasets}
