% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{get_topN_markers}
\alias{get_topN_markers}
\title{Get topN markers from each dataset}
\usage{
get_topN_markers(ref_list, reference, marker_top_n = 20)
}
\arguments{
\item{ref_list}{ref data list}

\item{reference}{'ectoM','mesoM','endoM','preM','exeM','ectoH','mesoH','endoH','preH','exeH'}

\item{marker_top_n}{default: 20}
}
\value{
topN markers from chosen dataset
}
\description{
Get topN markers from each dataset
}
