% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMarkers.R
\name{getMarkers}
\alias{getMarkers}
\title{getMarkers}
\usage{
getMarkers(species, reference, ref_celltype, ref_dataset, marker_top_n)
}
\arguments{
\item{species}{options: c('Hs','Mm')}

\item{reference}{options: c('ecto','meso','endo','exe','pre')}

\item{ref_celltype}{reference celltype}

\item{ref_dataset}{reference dataset}

\item{marker_top_n}{topN markers}
}
\value{
a list of markers
}
\description{
get markers from certain dataset and certain cell type
}
