% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMarkers.R
\name{getTotalMarkers}
\alias{getTotalMarkers}
\title{getTotalMarkers}
\usage{
getTotalMarkers(species, reference = "all")
}
\arguments{
\item{species}{options: c('Hs','Mm')}

\item{reference}{options: c('ecto','meso','endo','exe','pre')}
}
\value{
markers list
}
\description{
getTotalMarkers
}
