% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h.meso.pretrained_model}
\alias{h.meso.pretrained_model}
\title{pretrained model for human meso}
\format{
A data frame with gene rows and celltype cols
}
\usage{
h.meso.pretrained_model
}
\description{
pretrained model for human meso
}
\concept{data}
\keyword{datasets}
