% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h.pre.pretrained_model}
\alias{h.pre.pretrained_model}
\title{pretrained model for human pre}
\format{
A data frame with gene rows and celltype cols
}
\usage{
h.pre.pretrained_model
}
\description{
pretrained model for human pre
}
\concept{data}
\keyword{datasets}
