% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Heatmap_plot}
\alias{Heatmap_plot}
\title{Heatmap_plot}
\usage{
Heatmap_plot(
  auc,
  top_n = 3,
  use_shortname = T,
  color = colorRampPalette(rev(brewer.pal(n = 7, name = "RdBu")))(100),
  save = F,
  cutoff = 0.8,
  refine = T,
  custom.row=NULL,
  custom.col=NULL
)
}
\arguments{
\item{auc}{correlation bewtween query and ref celltypes (results from RUN_dscBLAST)}

\item{top_n}{top n correlated cell_type [default: 3]}

\item{use_shortname}{default: T}

\item{color}{vectors of color [default: colorRampPalette(rev(brewer.pal(n = 7, name = "RdBu")))(100)]}

\item{save}{default: F}

\item{refine}{default: T}

\item{custom.row}{If provided, filter rows in the auc matrix [default: NULL]}

\item{custom.col}{If provided, filter columns in the auc matrix [default: NULL]}
}
\description{
Heatmap plot
}
