% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMarkers.R
\name{loadMarkers}
\alias{loadMarkers}
\title{loadMarkers}
\usage{
loadMarkers(species, reference)
}
\arguments{
\item{species}{options: c('Hs','Mm')}

\item{reference}{options: c('ecto','meso','endo','exe','pre')}
}
\value{
markers list
}
\description{
load Markers from references
}
