% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m.ecto.pretrained_model}
\alias{m.ecto.pretrained_model}
\title{pretrained model for mouse ecto}
\format{
A data frame with gene rows and celltype cols
}
\usage{
m.ecto.pretrained_model
}
\description{
pretrained model for mouse ecto
}
\concept{data}
\keyword{datasets}
