% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m.meso.pretrained_model}
\alias{m.meso.pretrained_model}
\title{pretrained model for mouse meso}
\format{
A data frame with gene rows and celltype cols
}
\usage{
m.meso.pretrained_model
}
\description{
pretrained model for mouse meso
}
\concept{data}
\keyword{datasets}
