% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers_list_ectoH}
\alias{markers_list_ectoH}
\title{markerlist for human ecto}
\format{
A list of markers from different datasets
}
\usage{
markers_list_ectoH
}
\description{
markerlist for human ecto
}
\concept{data}
\keyword{datasets}
