% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers_list_ectoM}
\alias{markers_list_ectoM}
\title{markerlist for mouse ecto}
\format{
A list of markers from different datasets
}
\usage{
markers_list_ectoM
}
\description{
markerlist for mouse ecto
}
\concept{data}
\keyword{datasets}
