% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers_list_endoH}
\alias{markers_list_endoH}
\title{markerlist for human endo}
\format{
A list of markers from different datasets
}
\usage{
markers_list_endoH
}
\description{
markerlist for human endo
}
\concept{data}
\keyword{datasets}
