% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers_list_endoM}
\alias{markers_list_endoM}
\title{markerlist for mouse endo}
\format{
A list of markers from different datasets
}
\usage{
markers_list_endoM
}
\description{
markerlist for mouse endo
}
\concept{data}
\keyword{datasets}
