% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers_list_exeH}
\alias{markers_list_exeH}
\title{markerlist for human exe}
\format{
A list of markers from different datasets
}
\usage{
markers_list_exeH
}
\description{
markerlist for human exe
}
\concept{data}
\keyword{datasets}
