% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers_list_exeM}
\alias{markers_list_exeM}
\title{markerlist for mouse exe}
\format{
A list of markers from different datasets
}
\usage{
markers_list_exeM
}
\description{
markerlist for mouse exe
}
\concept{data}
\keyword{datasets}
