% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers_list_mesoH}
\alias{markers_list_mesoH}
\title{markerlist for human meso}
\format{
A list of markers from different datasets

A list of markers from different datasets
}
\usage{
markers_list_mesoH

markers_list_mesoH
}
\description{
markerlist for human meso

markerlist for human meso
}
\concept{data}
\keyword{datasets}
