% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers_list_preH}
\alias{markers_list_preH}
\title{markerlist for human pre}
\format{
A list of markers from different datasets
}
\usage{
markers_list_preH
}
\description{
markerlist for human pre
}
\concept{data}
\keyword{datasets}
