% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers_list_preM}
\alias{markers_list_preM}
\title{markerlist for mouse pre}
\format{
A list of markers from different datasets
}
\usage{
markers_list_preM
}
\description{
markerlist for mouse pre
}
\concept{data}
\keyword{datasets}
