% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotMarkers}
\alias{plotMarkers}
\title{Plot markers}
\usage{
plotMarkers(sce, species, features)
}
\arguments{
\item{sce}{a single cell experiment object}

\item{species}{options: c('Hs','Mm')}

\item{features}{genes offered for visualization}
}
\description{
Plot markers
}
