% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUN_dscBLAST.R
\name{RUN_dscBLAST.default}
\alias{RUN_dscBLAST.default}
\title{Title: RUN_dscBLAST.default}
\usage{
RUN_dscBLAST.default(sce, query_species, reference)
}
\arguments{
\item{sce}{a single cell experiment object}

\item{query_species}{options: c('Hs','Mm')}

\item{reference}{options: c('ecto','meso','endo','exe','pre')}
}
\value{
aurocs
correlation bewtween query and ref celltypes
}
\description{
Title: RUN_dscBLAST.default
}
