% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUN_dscBLAST_stage.R
\name{RUN_dscBLAST_stage}
\alias{RUN_dscBLAST_stage}
\title{RUN_dscBLAST_stage}
\usage{
RUN_dscBLAST_stage(
  sce,
  auc,
  query_species,
  top_n = 3,
  marker_top_n = 20,
  ref_dir
)
}
\arguments{
\item{sce}{a single cell experiment object}

\item{auc}{correlation bewtween query and ref celltypes (results from RUN_dscBLAST)}

\item{query_species}{options: c('Hs','Mm')}

\item{top_n}{default: 3}

\item{marker_top_n}{topN markers. N should be less than 20 [default: 20]}

\item{ref_dir}{ref data and meta stored here}
}
\value{
correlation between query and ref celltypes
}
\description{
run dscBLAST with stage info
}
