% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Sanky_plot}
\alias{Sanky_plot}
\title{Sanky_plot}
\usage{
Sanky_plot(
  auc,
  top_n = 3,
  color = c("#ffcc99", "#66cccc"),
  use_shortname = T,
  save = T,
  cutoff = 0.8,
  custom.row=NULL,
  custom.col=NULL
)
}
\arguments{
\item{auc}{correlation bewtween query and ref celltypes (results from RUN_dscBLAST)}

\item{top_n}{top n correlated cell_type [default: 3]}

\item{color}{vectors of colors [default: c("#ffcc99","#66cccc")]}

\item{use_shortname}{default: T}

\item{save}{default: T}

\item{custom.row}{If provided, filter rows in the auc matrix [default: NULL]}

\item{custom.col}{If provided, filter columns in the auc matrix [default: NULL]}
}

\description{
Sanky plot
}
