% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{subset_data}
\alias{subset_data}
\title{subset data}
\usage{
subset_data(auc, top_n, ref_dir)
}
\arguments{
\item{auc}{correlation bewtween query and ref celltypes(results from RUN_dscBLAST)}

\item{top_n}{top n correlated cell_type[default: 3]}

\item{ref_dir}{ref data and meta stored here}
}
\value{
a list of reference single cell experiment objects
}
\description{
subset data
}
