% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_marker.R
\name{subset_marker}
\alias{subset_marker}
\title{subset_marker}
\usage{
subset_marker(ref_list, marker_top_n)
}
\arguments{
\item{ref_list}{reference list}

\item{marker_top_n}{top N markers[default: 20]}
}
\value{
a list of ref markers
}
\description{
subset marker
}
