% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{substrRight}
\alias{substrRight}
\title{subRight character}
\usage{
substrRight(x, n = 1)
}
\arguments{
\item{x}{a character}

\item{n}{order from right\link{default: 1}}
}
\value{
sub_right one character
}
\description{
subRight character
}
