% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{trainModel_stage}
\alias{trainModel_stage}
\title{train model with stage info}
\usage{
trainModel_stage(var_genes, dat, i = 1, study_id, cell_type, stage_id)
}
\arguments{
\item{var_genes}{variable genes}

\item{dat}{a single cell experiment object}

\item{i}{a number\link{default:1}}

\item{study_id}{ref dataset info}

\item{cell_type}{ref celltype info}

\item{stage_id}{ref stage info}
}
\value{
a pretrained model with stage info
}
\description{
train model with stage info
}
