% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_gene.R
\name{transfer_gene}
\alias{transfer_gene}
\title{transfer_gene}
\usage{
transfer_gene(expression_profile, query_species)
}
\arguments{
\item{expression_profile}{a gene expression matrix}

\item{query_species}{options: c('Hs','Mm')}
}
\value{
a transfered cell-gene matrix
}
\description{
transfer mouse gene to human gene or in turn
}
