% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_calculate_withstage.R
\name{auc_calculate_withstage}
\alias{auc_calculate_withstage}
\title{auc_calculate_withstage}
\usage{
auc_calculate_withstage(sce, pretrained_model, query_species)
}
\arguments{
\item{sce}{a single cell experiment object}

\item{pretrained_model}{a pretrained model list}

\item{query_species}{options: c('Hs','Mm')}
}
\value{
correlation bewtween query and ref celltypes
}
\description{
calculate auc with stage info
}
