% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_qc.R
\name{Cell_qc}
\alias{Cell_qc}
\title{Cell_qc}
\usage{
Cell_qc(expression_profile, query_species, metadata, downsample = 50000)
}
\arguments{
\item{expression_profile}{a gene-cell expression matrix}

\item{query_species}{options: c('Hs','Mm')}

\item{metadata}{meta data containing info about input}
\item{cell_type}{cell type or cluster ID of input data}
\item{downsample}{set number of cells [default: 50000]}


}
\value{
a seurat object of cells passing qulity control
}
\description{
cell quality control
}
