% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainModel_stage.R
\name{design_matrix}
\alias{design_matrix}
\title{design_matrix}
\usage{
design_matrix(cell_type)
}
\arguments{
\item{cell_type}{ref celltype info}
}
\value{
a binary matix with celltype info
}
\description{
transform a vector with cell_type labels into a binary matrix
}
