% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topN_markers.R
\name{get_topN_markers}
\alias{get_topN_markers}
\title{get_topN_markers}
\usage{
get_topN_markers(ref_list, dataset.use, marker_top_n = 20)
}
\arguments{
\item{ref_list}{ref data list}

\item{reference}{options: c('ectoM','mesoM','endoM','preM','exeM','ectoH','mesoH','endoH','preH','exeH')}

\item{marker_top_n}{topN markers. N should be less than 20 [default: 20]}
}
\value{
topN markers from each dataset
}
\description{
get topN markers from each dataset
}
