% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{normalize_cols}
\alias{normalize_cols}
\title{Normalize matrix}
\usage{
normalize_cols(M, ranked = TRUE)
}
\arguments{
\item{M}{a matrix}

\item{ranked}{default: T}
}
\value{
a normalized matrix
}
\description{
Normalize matrix
}
