% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUN_dscBLAST.R
\name{RUN_dscBLAST}
\alias{RUN_dscBLAST}
\title{RUN_dscBLAST}
\usage{
RUN_dscBLAST(sce, query_species, reference = "all")
}
\arguments{
\item{sce}{a single cell experiment object}

\item{query_species}{options: c('Hs','Mm')}

\item{reference}{reference used. Generally we run dscblast in all datasets, and it takes time. If the origin of query cells is known and you prefer to run dscblast using certain reference, please change the option. options: c('all','ecto','meso','endo','exe','pre') [default: 'all'] }

\item{highlight_celltype}{celltypes to highlight [default: NULL]}
}
\value{
correlation bewtween query and ref celltypes
}
\description{
run dscBLAST
}
