% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{subset_marker}
\alias{subset_marker}
\title{Subset marker}
\usage{
subset_marker(ref_list, marker_top_n)
}
\arguments{
\item{ref_list}{reference list}

\item{marker_top_n}{top n markers
default: 20}
}
\value{
ref_marker_list
}
\description{
Subset marker
}
