% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_calculate_withstage.R
\name{substrRight}
\alias{substrRight}
\title{substrRight}
\usage{
substrRight(x, n=1)
}
\arguments{
\item{x}{a character}

\item{n}{order from right [default: 1]}
}
\value{
one character from right
}
\description{
subRight character
}
