% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{transfer_gene}
\alias{transfer_gene}
\title{Gene transfer}
\usage{
transfer_gene(expression_profile, query_species)
}
\arguments{
\item{expression_profile}{a gene expression matrix}

\item{query_species}{Hs/Mm}
}
\value{
a transfered gene expression matrix(human to mouse or mouse to human)
}
\description{
Gene transfer
}
