% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_gene.R
\name{transfer_gene}
\alias{transfer_gene}
\title{transfer_gene}
\usage{
  transfer_gene(expression_matrix, query_species)
}
\arguments{
  \item{expression_matrix}{a gene-cell matrix}
  
  \item{query_species}{options: c('Hs','Mm')}
}
\value{
  a transfered gene-cell matrix
}
\description{
  transfer mouse gene to human gene or in turn
}
